07-MNIST-Keras-TensorFlow/BackEnd/requirements.txt
absl-py==0.7.1
astor==0.8.0
cachetools==3.1.1
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.0.3
Flask-Cors==3.0.8
gast==0.2.2
google-auth==1.8.1
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.21.1
h5py==2.9.0
idna==2.8
imageio==2.5.0
imread==0.7.1
itsdangerous==1.1.0
Jinja2==2.10.1
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
numpy==1.17.4
oauthlib==3.1.0
opt-einsum==3.1.0
Pillow==6.2.1
protobuf==3.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
PyYAML==5.1.1
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.0.0
six==1.13.0
tensorboard==1.9.0
tensorflow>=1.15.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
urllib3==1.25.7
virtualenv==16.6.0
Werkzeug==0.15.4
wrapt==1.11.2